move manifest.py into ds_engine for better structure and fix tests and imports
