+++ src/virtual_stain_flow/datasets/ds_engine/manifest.py
@@ -26,6 +26,7 @@ import numpy as np
 import pandas as pd
 from PIL import Image
 
+
 @dataclass(frozen=True)
 class DatasetManifest:
     """
